Starting separate search context for api search
